requirements.txt
asgiref==3.9.2
black==25.9.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
cryptography==46.0.2
django==5.2.6
django-daraja==1.3.0
djangorestframework==3.16.1
dotenv==0.9.9
idna==3.10
mypy-extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.4.0
pycparser==2.23
python-decouple==3.8
python-dotenv==1.1.1
pytokens==0.1.10
requests==2.32.5
sqlparse==0.5.3
urllib3==2.5.0
